git-daemon will now compile on mingw32.  It is not functional at this point.
This commit makes the necessary changes to allow daemon.c to compile cleanly
on Mingw32.  We use the winsock2.h header and functions.  poll() does not
work properly so at this point it is compiled out.  There is not syslog
either, so that has been removed.

Currently, the daemon will serve one request and then quit.  You must start
the daemon using --base-path=<Windows path to root msys directory>.

Signed-off-by: Mike Pape <dotzenlabs@gmail.com>