Windows doesn't have fork so we work around using spawn and duplicating handles.
Without fork, we have to handle things in the current thread.  We use
spawnv_git_cmd to run the actual commands upload-pack, upload-archive,
and receive-pack.  Where it used to fork, we copy current descriptors,
execute the command with the incoming descriptor, close the copied
descriptors, and then copy the old descriptors back.  This allows the
daemon to process requests without exiting after the first one.

This still needs more work to be fully functional.

Signed-off-by: Mike Pape <dotzenlabs@gmail.com>