Teach core.autocrlf to 'git blame'
Pass the fake commit through convert_to_git, so that the
file is adjusted for local line-ending convention.

NOTE: This patch has already been sent upstream, to
git.git/next branch, and thus should only be considered
a temporary fix for the old code style which we're
currently on. (New code is based on recent strbuf work)