test-lib: Force core.autocrlf=false in test repository
The test scripts do not explicitly handle line endings but assume
that they are executed in a Unix environment with LF line
endings.  A global default core.autocrlf=true might confuse the
test scripts.  So better force the test repository to not convert
line endings.

Better would be if the test scripts worked with any core.autocrlf
setting.  However this would need more detailed investigations.
For now, this workaround should be sufficient.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>