WIP: unreliable hardlinks
It seems that accessing NTFS partitions with ufsd (at least on my EeePC)
has an unnerving bug: if you link() a file and unlink() it right away,
the target of the link() will have the correct size, but consist of
zeroes.

As ufsd is "Commertial software", I cannot fix it, and have to work
around it in Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>