Add diff option --ignore-untracked-in-submodules
Sometimes, submodules are _expected_ to contain untracked files. In that
case, the old behavior was desirable; even when adding untracked files
to .git/info/ignore in the submodules, they are traversed, I/O is
performed, disk caches are filled up, all of it unnecessarily.

In a very real case, this led to an increase of the runtime of "git status"
from unnoticable to 5 (in words: five) seconds, for no benefit at all.

This patch is the first building block to undo the damage.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>