copy content of config file into old
the old behaviour could result in problems if the config file is a symlink
because the symlinkfile will be replaced, not the real file.
i used the symlink to store all my config files at a central position.
see http://www.kerneltrap.com/mailarchive/git/2007/7/16/251763/thread
the fix is in the current version, but is not working for me.

Signed-off-by: Bastian Moldenhauer <bastianmoldenhauer@web.de>