Reintroduce symbolic links on windows
Use windows symbolic links (introduced in Vista) to mimic the behaviour
of symbolic links on other platforms. Based on an original patch made by
Thorvald Natvig.

This patch does not properly handle symbolic links to directories,
as directory links has to be flagged on NTFS for Windows to
properly recognice them. As the symbolic links are often created ahead of the
actual link targets, it is not possible to supply those extra flags
to the Windows API, why all symbolic links are created as file links.
File links to directories are handled by git itself, but native windows
tools might get problems.