add -e: include the full index lines to allow staging binary files' changes, too
This is a companion to the patch which taught 'git add -e' to offer only
submodule diffs that would actually apply. Because it does not make sense
to let the user edit a diff that does not apply without fixups in the
first place. Ever.

Unfortunately, just presenting the full index line is still not enough
when the postimage was never staged. But it's a start.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>