convert measure_time() to output in seconds using double
measure_time() currently prints and returns some cpu tick count. For
identifying spots which take a long time is sufficient to relatively
compare cpu ticks. In order to get some real world reference we
convert the output to seconds. This helps when trying to match
to the waited real time.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>