DRAFT: load builtins from internal list instead of directory
TODO: use get_internal_commands() in help.c

Previous to this patch the list of builtin commands were loaded from the
directory where hardlinks of the builtin commands to the main git
executable where stored. Not all platforms support hardlinks properly.
On these platforms (namely Windows) we might want to omit installing
them. This allows us to do so and still get a complete list of git
commands when skipping the hardlinkage of builtin binaries.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>