When opening a loose object failed with EINTR, try again
EINTR is an allowed error condition for the open() function and it
is non-fatal. So let's just try again if we hit that condition.

That condition is encountered quite frequently on this developer's
setup when reading objects from a disk mounted on MacOSX using
ntfs-3g (version 2013.1.13, via fuse4x 0.9.2).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>